k_order_perturbation: fix for compilation with MS Visual C++
git-svn-id: https://www.dynare.org/svn/dynare/trunk@3185 ac1d8469-bf42-47a9-8791-bf33cf982152

+++ mex/sources/k_order_perturbation/k_ord_dynare.cpp
@@ -249,9 +249,9 @@ KordpDynare::populateDerivativesContainer(TwoDMatrix *g, int ord, const vector<i
 	{
 	  int j = (int)g->get(i,0)-1; 
 	  int i1 = (int)g->get(i,1) -1;
-	  int s0 = (int)floor(i1/nJcols2);
+	  int s0 = (int)floor(((double) i1)/((double) nJcols2));
 	  int i2 = i1 - nJcols2*s0;
-	  int s1 = (int)floor(i2/nJcols);
+	  int s1 = (int)floor(((double) i2)/((double) nJcols));
 	  int s2 = i2 - nJcols*s1;
 	  if (s0 < nJcols1)
 	    s[0] = revOrder[s0];
